requirements.txt
# Requirements for ProteiNet + VAE + ESM-Fold Structure Tokenization

# Core PyTorch
torch>=2.0.0
torchvision>=0.15.0

# PyTorch Geometric and related packages
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
torch-cluster>=1.6.0

# PyTorch Lightning for training
pytorch-lightning>=2.0.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Data processing
pandas>=1.3.0
h5py>=3.7.0

# Visualization and logging
tensorboard>=2.10.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Protein structure handling
biopython>=1.79
prody>=2.0.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.64.0
pathlib2>=2.3.0

# Optional: For better performance
# Install these based on your CUDA version
# torch-scatter torch-sparse torch-cluster -f https://data.pyg.org/whl/torch-2.0.0+cu118.html

# Development dependencies
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0 